Extracts one base class
