+++ core/src/main/java/org/apache/struts2/dispatcher/multipart/JakartaStreamMultiPartRequest.java
@@ -1,8 +1,5 @@
 package org.apache.struts2.dispatcher.multipart;
 
-import com.opensymphony.xwork2.LocaleProvider;
-import com.opensymphony.xwork2.inject.Inject;
-import com.opensymphony.xwork2.util.LocalizedTextUtil;
 import org.apache.commons.fileupload.FileItemIterator;
 import org.apache.commons.fileupload.FileItemStream;
 import org.apache.commons.fileupload.FileUploadBase;
@@ -11,7 +8,6 @@ import org.apache.commons.fileupload.servlet.ServletFileUpload;
 import org.apache.commons.fileupload.util.Streams;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.struts2.StrutsConstants;
 import org.apache.struts2.dispatcher.LocalizedMessage;
 
 import javax.servlet.http.HttpServletRequest;
@@ -27,14 +23,9 @@ import java.util.*;
  * @author Chris Cranford
  * @since 2.3.18
  */
-public class JakartaStreamMultiPartRequest implements MultiPartRequest {
+public class JakartaStreamMultiPartRequest extends AbstractMultiPartRequest {
 
     static final Logger LOG = LogManager.getLogger(JakartaStreamMultiPartRequest.class);
-
-    /**
-     * Defines the internal buffer size used during streaming operations.
-     */
-    private static final int BUFFER_SIZE = 10240;
 
     /**
      * Map between file fields and file data.
@@ -45,55 +36,6 @@ public class JakartaStreamMultiPartRequest implements MultiPartRequest {
      * Map between non-file fields and values.
      */
     private Map<String, List<String>> parameters = new HashMap<>();
-
-    /**
-     * Internal list of raised errors to be passed to the the Struts2 framework.
-     */
-    private List<LocalizedMessage> errors = new ArrayList<>();
-
-    /**
-     * Internal list of non-critical messages to be passed to the Struts2 framework.
-     */
-    private List<String> messages = new ArrayList<>();
-
-    /**
-     * Specifies the maximum size of the entire request.
-     */
-    private Long maxSize;
-
-    /**
-     * Specifies the buffer size to use during streaming.
-     */
-    private int bufferSize = BUFFER_SIZE;
-
-    /**
-     * Localization to be used regarding errors.
-     */
-    private Locale defaultLocale = Locale.ENGLISH;
-
-    /**
-     * @param maxSize Injects the Struts multiple part maximum size.
-     */
-    @Inject(StrutsConstants.STRUTS_MULTIPART_MAXSIZE)
-    public void setMaxSize(String maxSize) {
-        this.maxSize = Long.parseLong(maxSize);
-    }
-
-    /**
-     * @param bufferSize Sets the buffer size to be used.
-     */
-    @Inject(value = StrutsConstants.STRUTS_MULTIPART_BUFFERSIZE, required = false)
-    public void setBufferSize(String bufferSize) {
-        this.bufferSize = Integer.parseInt(bufferSize);
-    }
-
-    /**
-     * @param provider Injects the Struts locale provider.
-     */
-    @Inject
-    public void setLocaleProvider(LocaleProvider provider) {
-        defaultLocale = provider.getLocale();
-    }
 
     /* (non-Javadoc)
      * @see org.apache.struts2.dispatcher.multipart.MultiPartRequest#cleanUp()
@@ -126,13 +68,6 @@ public class JakartaStreamMultiPartRequest implements MultiPartRequest {
         }
 
         return types.toArray(new String[types.size()]);
-    }
-
-    /* (non-Javadoc)
-     * @see org.apache.struts2.dispatcher.multipart.MultiPartRequest#getErrors()
-     */
-    public List<LocalizedMessage> getErrors() {
-        return errors;
     }
 
     /* (non-Javadoc)
@@ -235,16 +170,6 @@ public class JakartaStreamMultiPartRequest implements MultiPartRequest {
             if (!errors.contains(errorMessage)) {
                 errors.add(errorMessage);
             }
-        }
-    }
-
-    /**
-      * @param request Inspect the servlet request and set the locale if one wasn't provided by
-     * the Struts2 framework.
-     */
-    protected void setLocale(HttpServletRequest request) {
-        if (defaultLocale == null) {
-            defaultLocale = request.getLocale();
         }
     }
 
@@ -490,19 +415,6 @@ public class JakartaStreamMultiPartRequest implements MultiPartRequest {
             fileName = fileName.substring(backwardSlash + 1, fileName.length());
         }
         return fileName;
-    }
-
-    /**
-     * Build error message.
-     *
-     * @param e the Throwable/Exception
-     * @param args arguments
-     * @return error message
-     */
-    private LocalizedMessage buildErrorMessage(Throwable e, Object[] args) {
-        String errorKey = "struts.message.upload.error." + e.getClass().getSimpleName();
-        LOG.debug("Preparing error message for key: [{}]", errorKey);
-        return new LocalizedMessage(this.getClass(), errorKey, e.getMessage(), args);
     }
 
     /**
